requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.1.0
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.2.7
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
importlib-metadata==6.7.0
iniconfig==2.0.0
kiwisolver==1.4.4
mysql-connector-python==8.0.30
outcome==1.3.0.post0
packaging==21.3
Pillow==9.2.0
pluggy==1.2.0
protobuf==3.20.1
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.8.2
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
typing-extensions==4.3.0
urllib3==1.26.18
wsproto==1.2.0
zipp==3.15.0
